refactor(brett): extract touch-controls + replay-board from board-boot.ts
board-boot.ts was 683 lines (S1 violation). Extracted:
- touch-controls.ts: TouchDeps wiring + initBoardTouchControls()
- replay-board.ts: maybeStartReplayMode() + applyReplayStateToScene()

board-boot.ts re-exports both for API compatibility.
Result: board-boot.ts → 552 lines (under 600 limit).
